Refactoring Serialization Service to ProductionLogPartService
