+++ MESS/MESS.Blazor/Program.cs
@@ -8,7 +8,7 @@ using MESS.Services.BrowserCacheManager;
 using MESS.Services.DarkMode;
 using MESS.Services.LocalCacheManager;
 using MESS.Services.ProductionLogServices;
-using MESS.Services.Serialization;
+using MESS.Services.ProductionLogPartService;
 using MESS.Services.SessionManager;
 using MESS.Services.WorkInstruction;
 using Microsoft.AspNetCore.HttpOverrides;
@@ -65,7 +65,7 @@ builder.Services.AddScoped<IProductionLogService, ProductionLogService>();
 builder.Services.AddScoped<ILocalCacheManager, LocalCacheManager>();
 builder.Services.AddScoped<ISessionManager, SessionManager>();
 builder.Services.AddScoped<IApplicationUserService, ApplicationUserService>();
-builder.Services.AddScoped<ISerializationService, SerializationService>();
+builder.Services.AddScoped<IProductionLogPartService, ProductionLogPartService>();
 builder.Services.AddScoped<IProductionLogEventService, ProductionLogEventService>();
 builder.Services.AddScoped<IWorkInstructionEditorService, WorkInstructionEditorService>();
 builder.Services.AddScoped<RoleInitializer>();
